+++ samples/FluentMigrator.Net.Sdk/composite/Core/Migrations/M20260101000000_CreateCoreSchema.cs
@@ -1,0 +1,4 @@
+namespace Adventure.Core.Migrations
+{
+    public class M20260101000000_CreateCoreSchema { }
+}