+++ samples/FluentMigrator.Net.Sdk/composite/Sales/Migrations/M20260102000000_CreateOrders.cs
@@ -1,0 +1,4 @@
+namespace Adventure.Sales.Migrations
+{
+    public class M20260102000000_CreateOrders { }
+}